+++ lib/requests/diagnostics.go
