+++ lib/tracetracker/tracker.go
@@ -10,8 +10,10 @@ import (
 	"github.com/signalfx/golib/v3/datapoint"
 	"github.com/signalfx/golib/v3/sfxclient"
 	"github.com/signalfx/golib/v3/trace"
-	"github.com/signalfx/signalfx-agent/pkg/core/writer/correlations"
+	"github.com/signalfx/signalfx-agent/lib/correlations"
 	log "github.com/sirupsen/logrus"
+
+	"github.com/signalfx/signalfx-agent/pkg/core/writer/tracetracker"
 )
 
 const spanCorrelationMetricName = "sf.int.service.heartbeat"
@@ -191,8 +193,8 @@ func (a *ActiveServiceTracker) processEnvironment(span *trace.Span, now time.Tim
 		//
 		// Under that VERY specific circumstance, we need to fetch and delete the environment values for each
 		// pod/container that we have not already scraped an environment off of this agent runtime.
-		for i := range dimsToSyncSource {
-			dimName := dimsToSyncSource[i]
+		for i := range tracetracker.dimsToSyncSource {
+			dimName := tracetracker.dimsToSyncSource[i]
 			if dimValue, ok := span.Tags[dimName]; ok {
 				// look up the dimension / value in the environment cache to ensure it doesn't already exist
 				// if it does exist, this means we've already scraped and overwritten what was on the backend
@@ -265,7 +267,7 @@ func (a *ActiveServiceTracker) processEnvironment(span *trace.Span, now time.Tim
 
 	// container / pod level stuff
 	// this cache is necessary to identify environments associated with a kubernetes pod or container id
-	for _, dimName := range dimsToSyncSource {
+	for _, dimName := range tracetracker.dimsToSyncSource {
 		dimName := dimName
 		if dimValue, ok := span.Tags[dimName]; ok {
 			// Note that the value is not set on the cache key.  We only send the first environment received for a
@@ -329,7 +331,7 @@ func (a *ActiveServiceTracker) processService(span *trace.Span, now time.Time) {
 
 	// container / pod level stuff (this should not directly affect the active service count)
 	// this cache is necessary to identify services associated with a kubernetes pod or container id
-	for _, dimName := range dimsToSyncSource {
+	for _, dimName := range tracetracker.dimsToSyncSource {
 		dimName := dimName
 		if dimValue, ok := span.Tags[dimName]; ok {
 			// Note that the value is not set on the cache key.  We only send the first service received for a
